refactor: reduce the dependencies of CSS module
