+++ lib/platform/src/steptimer.c
