+++ test/xmake.lua
@@ -1,13 +1,12 @@
 includes("lib/test/xmake.lua")
+set_rundir("$(projectdir)/test")
+add_deps("LCUI")
 
 target("run_tests")
     set_default(false)
-    set_rundir("$(projectdir)/test")
     add_files("run_tests.c", "cases/*.c")
-    add_deps("LCUI", "test")
+    add_deps("test")
 
 target("helloworld")
     set_default(false)
-    set_rundir("$(projectdir)/test")
     add_files("helloworld.c")
-    add_deps("LCUI")
